# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HomeCare Platform API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  phone: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  partner: "partner",
  operator: "operator",
  admin: "admin",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  pending: "pending",
  suspended: "suspended",
  deleted: "deleted",
} as const;

export interface User {
  id: string;
  role: UserRole;
  name: string;
  phone: string;
  email?: string | null;
  status: UserStatus;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  token?: string;
}

export interface ServicePackage {
  id: string;
  name: string;
  slug: string;
  description: string;
  basePrice: number;
  estimatedMinutes: number;
  asWarrantyDays: number;
  isActive: boolean;
  sortOrder: number;
}

export type ServicePackageDetail = ServicePackage & {
  includedItems: string[];
  excludedItems: string[];
  tasks: string[];
};

export interface CreateOrderBody {
  packageId: string;
  roadAddress: string;
  detailAddress: string;
  scheduledDate: string;
  requestNote?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  requested: "requested",
  paid: "paid",
  pending_assignment: "pending_assignment",
  assigned: "assigned",
  en_route: "en_route",
  arrived: "arrived",
  in_progress: "in_progress",
  inspection_pending: "inspection_pending",
  inspection_approved: "inspection_approved",
  completed: "completed",
  cancelled: "cancelled",
  as_requested: "as_requested",
} as const;

export interface Order {
  id: string;
  orderNumber: string;
  customerId: string;
  packageId: string;
  packageName: string;
  status: OrderStatus;
  totalPrice: number;
  roadAddress: string;
  detailAddress: string;
  scheduledDate: string;
  requestNote?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderStatusLog {
  id: string;
  orderId: string;
  status: string;
  note?: string | null;
  createdAt: string;
}

export type JobAssignmentStatus =
  (typeof JobAssignmentStatus)[keyof typeof JobAssignmentStatus];

export const JobAssignmentStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface JobAssignment {
  id: string;
  orderId: string;
  partnerId: string;
  partnerName: string;
  partnerPhone: string;
  status: JobAssignmentStatus;
  scheduledDate: string;
  createdAt: string;
}

export interface Review {
  id: string;
  orderId: string;
  customerId: string;
  rating: number;
  comment?: string | null;
  createdAt: string;
}

export type OrderDetail = Order & {
  statusLogs: OrderStatusLog[];
  assignment?: JobAssignment | null;
  review?: Review | null;
};

export interface AdminOrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export type AdminOrderDetail = OrderDetail & {
  customerName: string;
  customerPhone: string;
  partnerName?: string | null;
  partnerPhone?: string | null;
};

export interface UpdateOrderStatusBody {
  status: string;
  note?: string | null;
}

export type AdminDashboardOrdersByStatusItem = {
  status: string;
  count: number;
};

export interface AdminDashboard {
  totalOrders: number;
  pendingAssignment: number;
  inProgress: number;
  inspectionPending: number;
  completedToday: number;
  asRequests: number;
  ordersByStatus: AdminDashboardOrdersByStatusItem[];
  recentOrders: Order[];
}

export interface AssignOrderBody {
  partnerId: string;
}

export type PartnerProfileBusinessType =
  (typeof PartnerProfileBusinessType)[keyof typeof PartnerProfileBusinessType];

export const PartnerProfileBusinessType = {
  individual: "individual",
  business: "business",
} as const;

export type PartnerProfileGrade =
  (typeof PartnerProfileGrade)[keyof typeof PartnerProfileGrade];

export const PartnerProfileGrade = {
  G1: "G1",
  G2: "G2",
  G3: "G3",
  G4: "G4",
  G5: "G5",
} as const;

export type PartnerProfileApprovalStatus =
  (typeof PartnerProfileApprovalStatus)[keyof typeof PartnerProfileApprovalStatus];

export const PartnerProfileApprovalStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  suspended: "suspended",
} as const;

export type PartnerProfileAvailabilityStatus =
  (typeof PartnerProfileAvailabilityStatus)[keyof typeof PartnerProfileAvailabilityStatus];

export const PartnerProfileAvailabilityStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface PartnerProfile {
  id: string;
  userId: string;
  name: string;
  phone: string;
  businessType: PartnerProfileBusinessType;
  grade: PartnerProfileGrade;
  ratingAvg: number;
  ratingCount: number;
  approvalStatus: PartnerProfileApprovalStatus;
  availabilityStatus: PartnerProfileAvailabilityStatus;
  createdAt: string;
}

export type ApprovePartnerBodyApprovalStatus =
  (typeof ApprovePartnerBodyApprovalStatus)[keyof typeof ApprovePartnerBodyApprovalStatus];

export const ApprovePartnerBodyApprovalStatus = {
  approved: "approved",
  rejected: "rejected",
  suspended: "suspended",
} as const;

export interface ApprovePartnerBody {
  approvalStatus: ApprovePartnerBodyApprovalStatus;
  note?: string | null;
}

export interface CreateReviewBody {
  orderId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}

export interface CreateAsRequestBody {
  orderId: string;
  reason: string;
  description?: string | null;
}

export type AsRequestStatus =
  (typeof AsRequestStatus)[keyof typeof AsRequestStatus];

export const AsRequestStatus = {
  pending: "pending",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface AsRequest {
  id: string;
  orderId: string;
  customerId: string;
  reason: string;
  description?: string | null;
  status: AsRequestStatus;
  createdAt: string;
}

export type ListOrdersParams = {
  status?: string;
};

export type AdminListOrdersParams = {
  status?: string;
  page?: number;
  limit?: number;
};

export type AdminListAssignmentsParams = {
  status?: string;
};

export type AdminListPartnersParams = {
  approvalStatus?: string;
};

export type AdminListAsRequestsParams = {
  status?: string;
};