# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HomeCare Platform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin list all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      summary: Admin get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail with assignment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
        "404":
          description: Not found
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/assignments:
    get:
      operationId: adminListAssignments
      tags: [admin]
      summary: List job assignments
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Assignment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobAssignment"
  /admin/orders/{id}/assign:
    post:
      operationId: adminAssignOrder
      tags: [admin]
      summary: Assign order to partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignOrderBody"
      responses:
        "200":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAssignment"
  /admin/partners:
    get:
      operationId: adminListPartners
      tags: [admin]
      summary: List all partners
      parameters:
        - name: approvalStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Partner list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerProfile"
  /admin/partners/{id}/approve:
    post:
      operationId: adminApprovePartner
      tags: [admin]
      summary: Approve or reject a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovePartnerBody"
      responses:
        "200":
          description: Partner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerProfile"
...
  /admin/as-requests:
    get:
      operationId: adminListAsRequests
      tags: [admin]
      summary: Admin list all A/S requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A/S requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AsRequest"
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [customer, partner, operator, admin]
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, suspended, deleted]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - name
        - phone
        - status
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerId:
          type: string
        packageId:
          type: string
        packageName:
          type: string
        status:
          type: string
          enum:
            - requested
            - paid
            - pending_assignment
            - assigned
            - en_route
            - arrived
            - in_progress
            - inspection_pending
            - inspection_approved
            - completed
            - cancelled
            - as_requested
        totalPrice:
          type: integer
        roadAddress:
          type: string
        detailAddress:
          type: string
        scheduledDate:
          type: string
          format: date-time
        requestNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerId
        - packageId
        - packageName
        - status
        - totalPrice
        - roadAddress
        - detailAddress
        - scheduledDate
        - createdAt
        - updatedAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            statusLogs:
              type: array
              items:
                $ref: "#/components/schemas/OrderStatusLog"
            assignment:
              $ref: "#/components/schemas/JobAssignment"
              nullable: true
            review:
              $ref: "#/components/schemas/Review"
              nullable: true
          required:
            - statusLogs
    OrderStatusLog:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        status:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - status
        - createdAt
    AdminOrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
    AdminOrderDetail:
      allOf:
        - $ref: "#/components/schemas/OrderDetail"
        - type: object
          properties:
            customerName:
              type: string
            customerPhone:
              type: string
            partnerName:
              type: string
              nullable: true
            partnerPhone:
              type: string
              nullable: true
          required:
            - customerName
            - customerPhone
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        note:
          type: string
          nullable: true
      required:
        - status
...
    JobAssignment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        partnerId:
          type: string
        partnerName:
          type: string
        partnerPhone:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled, completed]
        scheduledDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - partnerId
        - partnerName
        - partnerPhone
        - status
        - scheduledDate
        - createdAt
    AssignOrderBody:
      type: object
      properties:
        partnerId:
          type: string
      required:
        - partnerId
    PartnerProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: string
        businessType:
          type: string
          enum: [individual, business]
        grade:
          type: string
          enum: [G1, G2, G3, G4, G5]
        ratingAvg:
          type: number
        ratingCount:
          type: integer
        approvalStatus:
          type: string
          enum: [pending, approved, rejected, suspended]
        availabilityStatus:
          type: string
          enum: [available, busy, offline]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - phone
        - businessType
        - grade
        - ratingAvg
        - ratingCount
        - approvalStatus
        - availabilityStatus
        - createdAt
    ApprovePartnerBody:
      type: object
      properties:
        approvalStatus:
          type: string
          enum: [approved, rejected, suspended]
        note:
          type: string
          nullable: true
      required:
        - approvalStatus
...
    Review:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        customerId:
          type: string
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - customerId
        - rating
        - createdAt
...
    AsRequest:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        customerId:
          type: string
        reason:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - customerId
        - reason
        - status
        - createdAt